feat: adopt chart.js  (#607)
